Moved repositories to dedicated folders
