+++ src/Developurr.Orderly.Application/Command/Order/OpenOrder/OpenOrderUseCase.cs
@@ -1,6 +1,6 @@
-using Developurr.Orderly.Domain.Customer;
-using Developurr.Orderly.Domain.Order;
-using Developurr.Orderly.Domain.SalesConsultant;
+using Developurr.Orderly.Domain.Customer.Repositories;
+using Developurr.Orderly.Domain.Order.Repositories;
+using Developurr.Orderly.Domain.SalesConsultant.Repositories;
 
 namespace Developurr.Orderly.Application.Command.Order.OpenOrder;
