+++ src/Developurr.Orderly.Application/UseCase/Manager/GetManager/GetManagerUseCase.cs
@@ -1,5 +1,4 @@
-using Developurr.Orderly.Application.Command;
-using Developurr.Orderly.Domain.Manager;
+using Developurr.Orderly.Domain.Manager.Repositories;
 
 namespace Developurr.Orderly.Application.UseCase.Manager.GetManager;
