+++ src/Developurr.Orderly.Domain/Customer/Repositories/ICustomerRepository.cs
@@ -1,6 +1,6 @@
 using Developurr.Orderly.Domain.Customer.ValueObjects;
 using Developurr.Orderly.Domain.SeedWork;
 
-namespace Developurr.Orderly.Domain.Customer;
+namespace Developurr.Orderly.Domain.Customer.Repositories;
 
 public interface ICustomerRepository : IRepository<Customer, CustomerId> { }
