+++ src/Developurr.Orderly.Domain/Manager/Repositories/IManagerRepository.cs
@@ -1,6 +1,6 @@
 using Developurr.Orderly.Domain.Manager.ValueObjects;
 using Developurr.Orderly.Domain.SeedWork;
 
-namespace Developurr.Orderly.Domain.Manager;
+namespace Developurr.Orderly.Domain.Manager.Repositories;
 
 public interface IManagerRepository : IRepository<Manager, ManagerId> { }
