+++ src/Developurr.Orderly.Domain/Order/Repositories/IOrderRepository.cs
@@ -1,6 +1,6 @@
 using Developurr.Orderly.Domain.Order.ValueObjects;
 using Developurr.Orderly.Domain.SeedWork;
 
-namespace Developurr.Orderly.Domain.Order;
+namespace Developurr.Orderly.Domain.Order.Repositories;
 
 public interface IOrderRepository : IRepository<Order, OrderId> { }
