+++ src/Developurr.Orderly.Domain/Product/Repositories/IProductRepository.cs
@@ -1,6 +1,6 @@
 using Developurr.Orderly.Domain.Product.ValueObjects;
 using Developurr.Orderly.Domain.SeedWork;
 
-namespace Developurr.Orderly.Domain.Product;
+namespace Developurr.Orderly.Domain.Product.Repositories;
 
 public interface IProductRepository : IRepository<Product, ProductId> { }
