+++ src/Developurr.Orderly.Domain/SalesConsultant/Repositories/ISalesConsultantRepository.cs
@@ -1,6 +1,6 @@
 using Developurr.Orderly.Domain.SalesConsultant.ValueObjects;
 using Developurr.Orderly.Domain.SeedWork;
 
-namespace Developurr.Orderly.Domain.SalesConsultant;
+namespace Developurr.Orderly.Domain.SalesConsultant.Repositories;
 
 public interface ISalesConsultantRepository : IRepository<SalesConsultant, SalesConsultantId> { }
