+++ src/Developurr.Orderly.Domain/Shipping/Repositories/IShippingRepository.cs
@@ -1,6 +1,6 @@
 using Developurr.Orderly.Domain.SeedWork;
 using Developurr.Orderly.Domain.Shipping.ValueObjects;
 
-namespace Developurr.Orderly.Domain.Shipping;
+namespace Developurr.Orderly.Domain.Shipping.Repositories;
 
 public interface IShippingRepository : IRepository<Shipping, ShippingId> { }
