+++ tests/UnitTests/Developurr.Orderly.Application.UnitTests/TestUtils/CreateProduct/CreateProductFixture.cs
@@ -1,6 +1,6 @@
 using Developurr.Orderly.Application.Command;
 using Developurr.Orderly.Application.UseCase.Product.CreateProduct;
-using Developurr.Orderly.Domain.Product;
+using Developurr.Orderly.Domain.Product.Repositories;
 using Developurr.Orderly.Domain.UnitTests.TestUtils.Constants;
 using Moq;
