+++ tests/UnitTests/Developurr.Orderly.Application.UnitTests/TestUtils/GetProduct/GetProductFixture.cs
@@ -1,5 +1,5 @@
 using Developurr.Orderly.Application.UseCase.Product.GetProduct;
-using Developurr.Orderly.Domain.Product;
+using Developurr.Orderly.Domain.Product.Repositories;
 using Developurr.Orderly.Domain.UnitTests.TestUtils.Constants;
 using Moq;
